pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mypv2mqtt"
version = "0.1.0"
description = "Bridge my-PV ELWA data to MQTT topics."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE.md" }
dependencies = [
    "aiohttp",
    "paho-mqtt",
]

[project.scripts]
mypv2mqtt = "mypv2mqtt.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = []
